requirements.txt
fastapi
sqlalchemy
uvicorn[standard]
pydantic-settings
pymysql